Move and test bio image
