+++ src/components/bio.js
@@ -1,35 +1,8 @@
 import React from 'react';
 import { useStaticQuery, graphql } from 'gatsby';
-import Image from 'gatsby-image';
 import { rhythm } from '../utils/typography';
 import SocialLinks from './social-links';
-
-const BioImage = ({ image, alt }) => {
-  const styles = {
-    container: {
-      display: 'flex',
-      justifyContent: 'center',
-    },
-    component: {
-      marginBottom: rhythm(0.5),
-    },
-    img: {
-      borderRadius: '50%',
-      marginBottom: rhythm(0.5),
-    },
-  };
-
-  return (
-    <div css={styles.container}>
-      <Image
-        fixed={image}
-        alt={alt}
-        css={styles.component}
-        imgStyle={styles.img}
-      />
-    </div>
-  );
-};
+import BioImage from './bio-image';
 
 const Bio = () => {
   const data = useStaticQuery(graphql`
